// backend/src/domain/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  GERENTE = 'GERENTE',
  ATENDENTE = 'ATENDENTE',
  FINANCEIRO = 'FINANCEIRO',
  ESTOQUE = 'ESTOQUE',
  MECANICO = 'MECANICO',
  ADMINISTRADOR = 'ADMINISTRADOR',
  DIRETOR = 'DIRETOR',
  GESTOR_FINANCEIRO = 'GESTOR_FINANCEIRO',
  RECEPCIONISTA = 'RECEPCIONISTA',
  CONTADOR = 'CONTADOR',
}

export enum AuditLevel {
  BASICO = 'BASICO',
  INTERMEDIARIO = 'INTERMEDIARIO',
  AVANCADO = 'AVANCADO',
}

export enum TipoMovimentacao {
  ENTRADA = 'ENTRADA',
  ENTRADA_DEVOLUCAO_OS = 'ENTRADA_DEVOLUCAO_OS',
  SAIDA_OS = 'SAIDA_OS',
  SAIDA_PERDA = 'SAIDA_PERDA',
  SAIDA_PDV = 'SAIDA_PDV',
}

export enum StatusBarOS {
  ABERTA = 'ABERTA',
  EM_DIAGNOSTICO = 'EM_DIAGNOSTICO',
  AGUARDANDO_APROVACAO = 'AGUARDANDO_APROVACAO',
  APROVADA = 'APROVADA',
  AGUARDANDO_ASSINATURA = 'AGUARDANDO_ASSINATURA',
  ASSINADO = 'ASSINADO',
  EM_EXECUCAO = 'EM_EXECUCAO',
  AGUARDANDO_PECA = 'AGUARDANDO_PECA',
  CONCLUIDA = 'CONCLUIDA',
  ENTREGUE = 'ENTREGUE',
  CONCLUIDO = 'CONCLUIDO',
  CANCELADA = 'CANCELADA',
  PAGO = 'PAGO',
}

export enum TipoTransacao {
  RECEITA = 'RECEITA',
  DESPESA = 'DESPESA',
}

export enum MetodoPagamento {
  PIX = 'PIX',
  CREDITO = 'CREDITO',
  DEBITO = 'DEBITO',
  BOLETO = 'BOLETO',
  DINHEIRO = 'DINHEIRO',
}

export enum StatusFinanceiro {
  PENDENTE = 'PENDENTE',
  PAGO = 'PAGO',
  ATRASADO = 'ATRASADO',
}